removes un-needed fixture file adds error handler module for application
