--- cypress/fixtures/unsuccessfulRegistration.json
@@ -1,6 +1,0 @@
-{
-  "errors": [
-    "Something went wrong, please check the information you provided."
-  ],
-  "success": false
-}
