+++ src/modules/Articles.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import store from '../state/store/configureStore';
+import errorHandler from './ErrorHandler'
 
 const Articles = {
   async index(category) {
@@ -31,21 +32,3 @@ const Articles = {
 
 export default Articles;
 
-const errorHandler = (error) => {
-  if (error.response.status === 500) {
-    store.dispatch({
-      type: 'ERROR_MESSAGE',
-      payload: 'Servers are currently not responding, Please try again later',
-    });
-  } else if (error.response.status === 404) {
-    store.dispatch({
-      type: 'ERROR_MESSAGE',
-      payload: error.response.data.error_message,
-    });
-  } else {
-    store.dispatch({
-      type: 'ERROR_MESSAGE',
-      payload: error.message,
-    });
-  }
-};
